Revert "Add Guided Mode (#520)"
This reverts commit 88b9c993323645092abd635af75190f7d416b423.

Reimplementing on top of new analyzer.
