+++ Lib/Collectors/DriverCollector.cs
@@ -21,7 +21,7 @@ namespace AttackSurfaceAnalyzer.Collectors
         /// </summary>
         /// <param name="opts"> </param>
         /// <param name=""> </param>
-        public DriverCollector(CollectorOptions? opts = null, Action<CollectObject>? changeHandler = null) : base(opts, changeHandler) { }
+        public DriverCollector(CollectCommandOptions? opts = null, Action<CollectObject>? changeHandler = null) : base(opts, changeHandler) { }
 
         public override bool CanRunOnPlatform()
         {
