+++ Lib/Objects/FileMonitorObject.cs
@@ -4,13 +4,14 @@ using System.IO;
 
 namespace AttackSurfaceAnalyzer.Objects
 {
-    public class FileMonitorObject : MonitorObject
+    public class FileMonitorObject : CollectObject
     {
         public FileMonitorObject(string PathIn)
         {
             Path = PathIn;
         }
 
+        public CHANGE_TYPE? ChangeType { get; set; }
         public string? ExtendedResults { get; set; }
         public FileSystemObject? FileSystemObject { get; set; }
