--- Lib/Objects/MonitorObject.cs
@@ -1,15 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT License.
-using AttackSurfaceAnalyzer.Types;
-using AttackSurfaceAnalyzer.Utils;
-using System.Globalization;
-
-namespace AttackSurfaceAnalyzer.Objects
-{
-    /// <summary>
-    ///     Abstract parent class that all Collected data inherits from.
-    /// </summary>
-    public abstract class MonitorObject : CollectObject 
-    {
-        public CHANGE_TYPE ChangeType { get; set; }
-    }
-}
