+++ Tests/AnalyzerOperationsTests.cs
@@ -3,17 +3,47 @@ using AttackSurfaceAnalyzer.Objects;
 using AttackSurfaceAnalyzer.Types;
 using AttackSurfaceAnalyzer.Utils;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using Tpm2Lib;
-using Signature = AttackSurfaceAnalyzer.Objects.Signature;
 
 namespace AttackSurfaceAnalyzer.Tests
 {
     [TestClass]
-    public class AnalyzerOperationsTests
+    public class AnalyzerTests
     {
+        #region Private Fields
+
+        private const string TestPathOne = "TestPath1";
+        private const string TestPathTwo = "TestPath2";
+
+        private readonly CompareResult testPathOneExecutableObject = new CompareResult()
+        {
+            Base = new FileSystemObject(TestPathOne)
+            {
+                IsExecutable = true
+            }
+        };
+
+        private readonly CompareResult testPathOneObject = new CompareResult()
+        {
+            Base = new FileSystemObject(TestPathOne)
+        };
+
+        private readonly CompareResult testPathTwoExecutableObject = new CompareResult()
+        {
+            Base = new FileSystemObject(TestPathTwo)
+            {
+                IsExecutable = true
+            }
+        };
+
+        private readonly CompareResult testPathTwoObject = new CompareResult()
+        {
+            Base = new FileSystemObject(TestPathTwo)
+        };
+
+        #endregion Private Fields
+
         #region Public Methods
 
         [ClassInitialize]
@@ -21,1026 +51,141 @@ namespace AttackSurfaceAnalyzer.Tests
         {
             Logger.Setup(false, true);
             Strings.Setup();
-            AsaTelemetry.SetEnabled(enabled: false);
+            AsaTelemetry.SetEnabled(enabled:false);
         }
 
         [TestMethod]
-        public void VerifyContainsAnyOperator()
+        public void TestXorFromNand()
         {
-            var trueStringObject = new CompareResult()
+            var RuleName = "XOR from NAND";
+            var norRule = new Rule(RuleName)
             {
-                Base = new FileSystemObject("ContainsStringObject")
-            };
-
-            var alsoTrueStringObject = new CompareResult()
-            {
-                Base = new FileSystemObject("StringObject")
-            };
-
-            var falseStringObject = new CompareResult()
-            {
-                Base = new FileSystemObject("NothingInCommon")
-            };
-
-            var stringContains = new Rule("String Contains Any Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Path", OPERATION.CONTAINS_ANY)
-                    {
-                        Data = new List<string>()
-                        {
-                            "String",
-                        }
-                    }
-                }
-            };
-
-            var stringAnalyzer = GetAnalyzerForRule(stringContains);
-
-            Assert.IsTrue(stringAnalyzer.Analyze(trueStringObject).Any());
-            Assert.IsTrue(stringAnalyzer.Analyze(alsoTrueStringObject).Any());
-            Assert.IsFalse(stringAnalyzer.Analyze(falseStringObject).Any());
-
-            var trueListObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Subkeys = new List<string>()
-                    {
-                        "One",
-                        "Two",
-                        "Three"
-                    }
-                }
-            };
-
-            var alsoTrueListObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Subkeys = new List<string>()
-                    {
-                        "One",
-                        "Two",
-                    }
-                }
-            };
-
-            var falseListObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
-            };
-
-            var listContains = new Rule("List Contains Any Rule")
-            {
-                ResultType = RESULT_TYPE.REGISTRY,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Subkeys", OPERATION.CONTAINS_ANY)
-                    {
-                        Data = new List<string>()
-                        {
-                            "One",
-                            "Two",
-                            "Three"
-                        }
-                    }
-                }
-            };
-
-            var listAnalyzer = GetAnalyzerForRule(listContains);
-
-            Assert.IsTrue(listAnalyzer.Analyze(trueListObject).Any());
-            Assert.IsTrue(listAnalyzer.Analyze(alsoTrueListObject).Any());
-            Assert.IsFalse(listAnalyzer.Analyze(falseListObject).Any());
-
-            var trueStringDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsStringDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Values = new Dictionary<string, string>()
-                    {
-                        { "One", "One" },
-                        { "Two", "Two" },
-                        { "Three", "Three" }
-                    }
-                }
-            };
-
-            var alsoTrueStringDict = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsStringDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Values = new Dictionary<string, string>()
-                    {
-                        { "One", "One" },
-                        { "Two", "Three" },
-                    }
-                }
-            };
-
-            var superFalseStringDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsStringDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Values = new Dictionary<string, string>()
-                    {
-                        { "One", "Two" },
-                        { "Three", "Four" },
-                    }
-                }
-            };
-
-            var stringDictContains = new Rule("String Dict Contains Any Rule")
-            {
-                ResultType = RESULT_TYPE.REGISTRY,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Values", OPERATION.CONTAINS_ANY)
-                    {
-                        DictData = new List<KeyValuePair<string, string>>()
-                        {
-                            new KeyValuePair<string, string>("One","One"),
-                            new KeyValuePair<string, string>("Two","Two"),
-                            new KeyValuePair<string, string>("Three","Three")
-                        }
-                    }
-                }
-            };
-
-            var stringDictAnalyzer = GetAnalyzerForRule(stringDictContains);
-
-            Assert.IsTrue(stringDictAnalyzer.Analyze(trueStringDictObject).Any());
-            Assert.IsTrue(stringDictAnalyzer.Analyze(alsoTrueStringDict).Any());
-            Assert.IsFalse(stringDictAnalyzer.Analyze(superFalseStringDictObject).Any());
-
-            var trueListDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Permissions = new Dictionary<string, List<string>>()
-                    {
-                        {
-                            "User", new List<string>()
-                            {
-                                "Read",
-                                "Execute"
-                            }
-                        }
-                    }
-                }
-            };
-
-            var alsoTrueListDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Permissions = new Dictionary<string, List<string>>()
-                    {
-                        {
-                            "User", new List<string>()
-                            {
-                                "Read",
-                            }
-                        }
-                    }
-                }
-            };
-
-            var falseListDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Permissions = new Dictionary<string, List<string>>()
-                    {
-                        {
-                            "Taco", new List<string>()
-                            {
-                                "Read",
-                                "Execute"
-                            }
-                        }
-                    }
-                }
-            };
-
-            var listDictContains = new Rule("List Dict Contains Any Rule")
-            {
-                ResultType = RESULT_TYPE.REGISTRY,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Permissions", OPERATION.CONTAINS_ANY)
-                    {
-                        DictData = new List<KeyValuePair<string, string>>()
-                        {
-                            new KeyValuePair<string, string>("User","Execute"),
-                            new KeyValuePair<string, string>("User","Read")
-                        }
-                    }
-                }
-            };
-
-            var listDictAnalyzer = GetAnalyzerForRule(listDictContains);
-
-            Assert.IsTrue(listDictAnalyzer.Analyze(trueListDictObject).Any());
-            Assert.IsTrue(listDictAnalyzer.Analyze(alsoTrueListDictObject).Any());
-            Assert.IsFalse(listDictAnalyzer.Analyze(falseListDictObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyContainsKeyOperator()
-        {
-            var trueAlgDict = new CompareResult()
-            {
-                Base = new TpmObject("TestLocal")
-                {
-                    PCRs = new Dictionary<(Tpm2Lib.TpmAlgId, uint), byte[]>()
-                    {
-                        { (TpmAlgId.Sha,1), Array.Empty<byte>() }
-                    }
-                }
-            };
-
-            var falseAlgDict = new CompareResult()
-            {
-                Base = new TpmObject("TestLocal")
-                {
-                    PCRs = new Dictionary<(Tpm2Lib.TpmAlgId, uint), byte[]>()
-                    {
-                        { (TpmAlgId.Sha,15), Array.Empty<byte>() }
-                    }
-                }
-            };
-
-            var algDictContains = new Rule("Alg Dict Changed PCR 1")
-            {
-                ResultType = RESULT_TYPE.TPM,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("PCRs", OPERATION.CONTAINS_KEY)
-                    {
-                        Data = new List<string>()
-                        {
-                            "(Sha, 1)"
-                        }
-                    }
-                }
-            };
-
-            var algDictAnalyzer = GetAnalyzerForRule(algDictContains);
-
-            Assert.IsTrue(algDictAnalyzer.Analyze(trueAlgDict).Any());
-            Assert.IsFalse(algDictAnalyzer.Analyze(falseAlgDict).Any());
-        }
-
-        [TestMethod]
-        public void VerifyContainsOperator()
-        {
-            var trueStringObject = new CompareResult()
-            {
-                Base = new FileSystemObject("ContainsStringObject")
-            };
-
-            var falseStringObject = new CompareResult()
-            {
-                Base = new FileSystemObject("StringObject")
-            };
-
-            var superFalseStringObject = new CompareResult()
-            {
-                Base = new FileSystemObject("NothingInCommon")
-            };
-
-            var stringContains = new Rule("String Contains Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Path", OPERATION.CONTAINS)
-                    {
-                        Data = new List<string>()
-                        {
-                            "Contains",
-                            "String",
-                            "Object"
-                        }
-                    }
-                }
-            };
-
-            var stringAnalyzer = GetAnalyzerForRule(stringContains);
-
-            Assert.IsTrue(stringAnalyzer.Analyze(trueStringObject).Any());
-            Assert.IsFalse(stringAnalyzer.Analyze(falseStringObject).Any());
-            Assert.IsFalse(stringAnalyzer.Analyze(superFalseStringObject).Any());
-
-            var trueListObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Subkeys = new List<string>()
-                    {
-                        "One",
-                        "Two",
-                        "Three"
-                    }
-                }
-            };
-
-            var falseListObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Subkeys = new List<string>()
-                    {
-                        "One",
-                        "Two",
-                    }
-                }
-            };
-
-            var superFalseListObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
-            };
-
-            var listContains = new Rule("List Contains Rule")
-            {
-                ResultType = RESULT_TYPE.REGISTRY,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Subkeys", OPERATION.CONTAINS)
-                    {
-                        Data = new List<string>()
-                        {
-                            "One",
-                            "Two",
-                            "Three"
-                        }
-                    }
-                }
-            };
-
-            var listAnalyzer = GetAnalyzerForRule(listContains);
-
-            Assert.IsTrue(listAnalyzer.Analyze(trueListObject).Any());
-            Assert.IsFalse(listAnalyzer.Analyze(falseListObject).Any());
-            Assert.IsFalse(listAnalyzer.Analyze(superFalseListObject).Any());
-
-            var trueStringDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsStringDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Values = new Dictionary<string, string>()
-                    {
-                        { "One", "One" },
-                        { "Two", "Two" },
-                        { "Three", "Three" }
-                    }
-                }
-            };
-
-            var falseStringDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsStringDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Values = new Dictionary<string, string>()
-                    {
-                        { "One", "One" },
-                        { "Two", "Three" },
-                    }
-                }
-            };
-
-            var superFalseStringDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsStringDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Values = new Dictionary<string, string>()
-                    {
-                        { "One", "Two" },
-                        { "Three", "Four" },
-                    }
-                }
-            };
-
-            var stringDictContains = new Rule("String Dict Contains Rule")
-            {
-                ResultType = RESULT_TYPE.REGISTRY,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Values", OPERATION.CONTAINS)
-                    {
-                        DictData = new List<KeyValuePair<string, string>>()
-                        {
-                            new KeyValuePair<string, string>("One","One"),
-                            new KeyValuePair<string, string>("Two","Two"),
-                            new KeyValuePair<string, string>("Three","Three")
-                        }
-                    }
-                }
-            };
-
-            var stringDictAnalyzer = GetAnalyzerForRule(stringDictContains);
-
-            Assert.IsTrue(stringDictAnalyzer.Analyze(trueStringDictObject).Any());
-            Assert.IsFalse(stringDictAnalyzer.Analyze(falseStringDictObject).Any());
-            Assert.IsFalse(stringDictAnalyzer.Analyze(superFalseStringDictObject).Any());
-
-            var trueListDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Permissions = new Dictionary<string, List<string>>()
-                    {
-                        {
-                            "User", new List<string>()
-                            {
-                                "Read",
-                                "Execute"
-                            }
-                        }
-                    }
-                }
-            };
-
-            var falseListDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Permissions = new Dictionary<string, List<string>>()
-                    {
-                        {
-                            "User", new List<string>()
-                            {
-                                "Read",
-                            }
-                        }
-                    }
-                }
-            };
-
-            var alsoFalseListDictObject = new CompareResult()
-            {
-                Base = new RegistryObject("ContainsListDictObject", Microsoft.Win32.RegistryView.Registry32)
-                {
-                    Permissions = new Dictionary<string, List<string>>()
-                    {
-                        {
-                            "Contoso", new List<string>()
-                            {
-                                "Read",
-                                "Execute"
-                            }
-                        }
-                    }
-                }
-            };
-
-            var listDictContains = new Rule("List Dict Contains Rule")
-            {
-                ResultType = RESULT_TYPE.REGISTRY,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Permissions", OPERATION.CONTAINS)
-                    {
-                        DictData = new List<KeyValuePair<string, string>>()
-                        {
-                            new KeyValuePair<string, string>("User","Execute"),
-                            new KeyValuePair<string, string>("User","Read"),
-                        }
-                    }
-                }
-            };
-
-            var listDictAnalyzer = GetAnalyzerForRule(listDictContains);
-
-            Assert.IsTrue(listDictAnalyzer.Analyze(trueListDictObject).Any());
-            Assert.IsFalse(listDictAnalyzer.Analyze(falseListDictObject).Any());
-            Assert.IsFalse(listDictAnalyzer.Analyze(alsoFalseListDictObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyEndsWithOperator()
-        {
-            var trueEndsWithObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-            };
-            var falseEndsWithObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.pdf")
-            };
-
-            var endsWithRule = new Rule("Ends With Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Path", OPERATION.ENDS_WITH)
-                    {
-                        Data = new List<string>()
-                        {
-                            ".exe"
-                        }
-                    }
-                }
-            };
-
-            var endsWithAnalyzer = GetAnalyzerForRule(endsWithRule);
-
-            Assert.IsTrue(endsWithAnalyzer.Analyze(trueEndsWithObject).Any());
-            Assert.IsFalse(endsWithAnalyzer.Analyze(falseEndsWithObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyEqOperator()
-        {
-            var assertTrueObject = new CompareResult()
-            {
-                Base = new FileSystemObject("TestPath")
-                {
-                    IsDirectory = true,
-                    Size = 700
-                }
-            };
-
-            var assertFalseObject = new CompareResult()
-            {
-                Base = new FileSystemObject("TestPath2")
-                {
-                    IsDirectory = false,
-                    Size = 701
-                }
-            };
-
-            var stringEquals = new Rule("String Equals Rule")
-            {
+                Expression = "(0 NAND (0 NAND 1)) NAND (1 NAND (0 NAND 1))",
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
                     new Clause("Path", OPERATION.EQ)
                     {
+                        Label = "0",
                         Data = new List<string>()
                         {
-                            "TestPath"
+                            "TestPath1"
                         }
-                    }
-                }
-            };
-
-            var boolEquals = new Rule("Bool Equals Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("IsDirectory", OPERATION.EQ)
-                    {
-                        Data = new List<string>()
-                        {
-                            "True"
-                        }
-                    }
-                }
-            };
-
-            var intEquals = new Rule("Int Equals Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Size", OPERATION.EQ)
-                    {
-                        Data = new List<string>()
-                        {
-                            "700"
-                        }
-                    }
-                }
-            };
-
-            var boolAnalyzer = GetAnalyzerForRule(boolEquals);
-            var intAnalyzer = GetAnalyzerForRule(intEquals);
-            var stringAnalyzer = GetAnalyzerForRule(stringEquals);
-
-            Assert.IsTrue(boolAnalyzer.Analyze(assertTrueObject).Any(x => x.Name == "Bool Equals Rule"));
-            Assert.IsTrue(intAnalyzer.Analyze(assertTrueObject).Any(x => x.Name == "Int Equals Rule"));
-            Assert.IsTrue(stringAnalyzer.Analyze(assertTrueObject).Any(x => x.Name == "String Equals Rule"));
-
-            Assert.IsFalse(boolAnalyzer.Analyze(assertFalseObject).Any(x => x.Name == "Bool Equals Rule"));
-            Assert.IsFalse(intAnalyzer.Analyze(assertFalseObject).Any(x => x.Name == "Int Equals Rule"));
-            Assert.IsFalse(stringAnalyzer.Analyze(assertFalseObject).Any(x => x.Name == "String Equals Rule"));
-        }
-
-        [TestMethod]
-        public void VerifyGtOperator()
-        {
-            var trueGtObject = new CompareResult()
-            {
-                Base = new OpenPortObject(1025, TRANSPORT.TCP, ADDRESS_FAMILY.InterNetwork)
-            };
-            var falseGtObject = new CompareResult()
-            {
-                Base = new OpenPortObject(1023, TRANSPORT.TCP, ADDRESS_FAMILY.InterNetwork)
-            };
-
-            var gtRule = new Rule("Gt Rule")
-            {
-                ResultType = RESULT_TYPE.PORT,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Port", OPERATION.GT)
-                    {
-                        Data = new List<string>()
-                        {
-                            "1024"
-                        }
-                    }
-                }
-            };
-
-            var badGtRule = new Rule("Bad Gt Rule")
-            {
-                ResultType = RESULT_TYPE.PORT,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Port", OPERATION.GT)
-                    {
-                        Data = new List<string>()
-                        {
-                            "CONTOSO"
-                        }
-                    }
-                }
-            };
-
-            var gtAnalyzer = GetAnalyzerForRule(gtRule);
-
-            Assert.IsTrue(gtAnalyzer.Analyze(trueGtObject).Any());
-            Assert.IsFalse(gtAnalyzer.Analyze(falseGtObject).Any());
-
-            var badGtAnalyzer = GetAnalyzerForRule(badGtRule);
-
-            Assert.IsFalse(badGtAnalyzer.Analyze(trueGtObject).Any());
-            Assert.IsFalse(badGtAnalyzer.Analyze(falseGtObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyIsAfterOperator()
-        {
-            var falseIsAfterObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    SignatureStatus = new Signature(true)
-                    {
-                        SigningCertificate = new SerializableCertificate(Thumbprint: string.Empty, Subject: string.Empty, PublicKey: string.Empty, NotAfter: DateTime.Now, NotBefore: DateTime.Now, Issuer: string.Empty, SerialNumber: string.Empty, CertHashString: string.Empty, Pkcs7: string.Empty)
-                    }
-                }
-            };
-
-            var trueIsAfterObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    SignatureStatus = new Signature(true)
-                    {
-                        SigningCertificate = new SerializableCertificate(Thumbprint: string.Empty, Subject: string.Empty, PublicKey: string.Empty, NotAfter: DateTime.Now.AddYears(1), NotBefore: DateTime.Now, Issuer: string.Empty, SerialNumber: string.Empty, CertHashString: string.Empty, Pkcs7: string.Empty)
-                    }
-                }
-            };
-
-            var isAfterRule = new Rule("Is After Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("SignatureStatus.SigningCertificate.NotAfter", OPERATION.IS_AFTER)
-                    {
-                        Data = new List<string>()
-                        {
-                            DateTime.Now.AddDays(1).ToString()
-                        }
-                    }
-                }
-            };
-
-            var isAfterAnalyzer = GetAnalyzerForRule(isAfterRule);
-
-            Assert.IsTrue(isAfterAnalyzer.Analyze(trueIsAfterObject).Any());
-            Assert.IsFalse(isAfterAnalyzer.Analyze(falseIsAfterObject).Any());
-
-            var isAfterRuleShortDate = new Rule("Is After Short Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("SignatureStatus.SigningCertificate.NotAfter", OPERATION.IS_AFTER)
-                    {
-                        Data = new List<string>()
-                        {
-                            DateTime.Now.AddDays(1).ToShortDateString()
-                        }
-                    }
-                }
-            };
-
-            var isAfterShortAnalyzer = GetAnalyzerForRule(isAfterRuleShortDate);
-
-            Assert.IsTrue(isAfterShortAnalyzer.Analyze(trueIsAfterObject).Any());
-            Assert.IsFalse(isAfterShortAnalyzer.Analyze(falseIsAfterObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyIsBeforeOperator()
-        {
-            var trueIsBeforeObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    SignatureStatus = new Signature(true)
-                    {
-                        SigningCertificate = new SerializableCertificate(Thumbprint: string.Empty, Subject: string.Empty, PublicKey: string.Empty, NotAfter: DateTime.Now, NotBefore: DateTime.Now, Issuer: string.Empty, SerialNumber: string.Empty, CertHashString: string.Empty, Pkcs7: string.Empty)
-                    }
-                }
-            };
-
-            var falseIsBeforeObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    SignatureStatus = new Signature(true)
-                    {
-                        SigningCertificate = new SerializableCertificate(Thumbprint: string.Empty, Subject: string.Empty, PublicKey: string.Empty, NotAfter: DateTime.Now.AddYears(1), NotBefore: DateTime.Now, Issuer: string.Empty, SerialNumber: string.Empty, CertHashString: string.Empty, Pkcs7: string.Empty)
-                    }
-                }
-            };
-
-            var isBeforeRule = new Rule("Is Before Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("SignatureStatus.SigningCertificate.NotAfter", OPERATION.IS_BEFORE)
-                    {
-                        Data = new List<string>()
-                        {
-                            DateTime.Now.AddDays(1).ToString()
-                        }
-                    }
-                }
-            };
-
-            var isBeforeAnalyzer = GetAnalyzerForRule(isBeforeRule);
-
-            Assert.IsTrue(isBeforeAnalyzer.Analyze(trueIsBeforeObject).Any());
-            Assert.IsFalse(isBeforeAnalyzer.Analyze(falseIsBeforeObject).Any());
-
-            var isBeforeShortRule = new Rule("Is Before Short Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("SignatureStatus.SigningCertificate.NotAfter", OPERATION.IS_BEFORE)
-                    {
-                        Data = new List<string>()
-                        {
-                            DateTime.Now.AddDays(1).ToShortDateString()
-                        }
-                    }
-                }
-            };
-
-            var isBeforeShortAnalyzer = GetAnalyzerForRule(isBeforeShortRule);
-
-            Assert.IsTrue(isBeforeShortAnalyzer.Analyze(trueIsBeforeObject).Any());
-            Assert.IsFalse(isBeforeShortAnalyzer.Analyze(falseIsBeforeObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyIsExpiredOperator()
-        {
-            var trueIsExpiredObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    SignatureStatus = new Signature(true)
-                    {
-                        SigningCertificate = new SerializableCertificate(Thumbprint: string.Empty, Subject: string.Empty, PublicKey: string.Empty, NotAfter: DateTime.MinValue, NotBefore: DateTime.Now, Issuer: string.Empty, SerialNumber: string.Empty, CertHashString: string.Empty, Pkcs7: string.Empty)
-                    }
-                }
-            };
-
-            var falseIsExpiredObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    SignatureStatus = new Signature(true)
-                    {
-                        SigningCertificate = new SerializableCertificate(Thumbprint: string.Empty, Subject: string.Empty, PublicKey: string.Empty, NotAfter: DateTime.MaxValue, NotBefore: DateTime.Now, Issuer: string.Empty, SerialNumber: string.Empty, CertHashString: string.Empty, Pkcs7: string.Empty)
-                    }
-                }
-            };
-
-            var isExpiredRule = new Rule("Is Expired Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("SignatureStatus.SigningCertificate.NotAfter", OPERATION.IS_EXPIRED)
-                }
-            };
-
-            var isExpiredAnalyzer = GetAnalyzerForRule(isExpiredRule);
-
-            Assert.IsTrue(isExpiredAnalyzer.Analyze(trueIsExpiredObject).Any());
-            Assert.IsFalse(isExpiredAnalyzer.Analyze(falseIsExpiredObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyIsNullOperator()
-        {
-            var falseIsNullObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    ContentHash = "HASH"
-                }
-            };
-            var trueIsNullObject = new CompareResult()
-            {
-                Base = new FileSystemObject("NotAnApp.pdf")
-            };
-
-            var isNullRule = new Rule("Is Null Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("ContentHash", OPERATION.IS_NULL)
-                }
-            };
-
-            var isNullAnalyzer = GetAnalyzerForRule(isNullRule);
-
-            Assert.IsTrue(isNullAnalyzer.Analyze(trueIsNullObject).Any());
-            Assert.IsFalse(isNullAnalyzer.Analyze(falseIsNullObject).Any());
-        }
-
-        [TestMethod]
-        public void VerifyIsTrueOperator()
-        {
-            var trueIsTrueObject = new CompareResult()
-            {
-                Base = new FileSystemObject("App.exe")
-                {
-                    IsExecutable = true
-                }
-            };
-            var falseIsTrueObject = new CompareResult()
-            {
-                Base = new FileSystemObject("NotAnApp.pdf")
-            };
-
-            var isTrueRule = new Rule("Is True Rule")
-            {
-                ResultType = RESULT_TYPE.FILE,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
+                    },
                     new Clause("IsExecutable", OPERATION.IS_TRUE)
+                    {
+                        Label = "1"
+                    }
                 }
             };
 
-            var isTrueAnalyzer = GetAnalyzerForRule(isTrueRule);
+            var analyzer = GetAnalyzerForRule(norRule);
 
-            Assert.IsTrue(isTrueAnalyzer.Analyze(trueIsTrueObject).Any());
-            Assert.IsFalse(isTrueAnalyzer.Analyze(falseIsTrueObject).Any());
+            Assert.IsTrue(analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathOneExecutableObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoExecutableObject).Any(x => x.Name == RuleName));
         }
 
         [TestMethod]
-        public void VerifyLtOperator()
+        public void VerifyAccessSubproperties()
         {
-            var falseLtObject = new CompareResult()
+            var regObj = new CompareResult()
             {
-                Base = new OpenPortObject(1025, TRANSPORT.TCP, ADDRESS_FAMILY.InterNetwork)
-            };
-            var trueLtObject = new CompareResult()
-            {
-                Base = new OpenPortObject(1023, TRANSPORT.TCP, ADDRESS_FAMILY.InterNetwork)
+                Base = new RegistryObject("ContainsListObject", Microsoft.Win32.RegistryView.Registry32)
+                {
+                    Values = new Dictionary<string, string>()
+                    {
+                        { "One", "Two"}
+                    }
+                }
             };
 
-            var ltRule = new Rule("Lt Rule")
+            var RuleName = "ContainsRule";
+            var containsRule = new Rule(RuleName)
             {
-                ResultType = RESULT_TYPE.PORT,
+                ResultType = RESULT_TYPE.REGISTRY,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("Port", OPERATION.LT)
+                    new Clause("Values.One", OPERATION.EQ)
                     {
+                        Label = "0",
                         Data = new List<string>()
                         {
-                            "1024"
+                            "Two"
                         }
                     }
                 }
             };
 
-            var ltAnalyzer = GetAnalyzerForRule(ltRule);
-
-            Assert.IsTrue(ltAnalyzer.Analyze(trueLtObject).Any());
-            Assert.IsFalse(ltAnalyzer.Analyze(falseLtObject).Any());
-
-            var badLtRule = new Rule("Bad Lt Rule")
-            {
-                ResultType = RESULT_TYPE.PORT,
-                Flag = ANALYSIS_RESULT_TYPE.FATAL,
-                Clauses = new List<Clause>()
-                {
-                    new Clause("Port", OPERATION.GT)
-                    {
-                        Data = new List<string>()
-                        {
-                            "CONTOSO"
-                        }
-                    }
-                }
-            };
-
-            var badLtAnalyzer = GetAnalyzerForRule(badLtRule);
-
-            Assert.IsFalse(badLtAnalyzer.Analyze(trueLtObject).Any());
-            Assert.IsFalse(badLtAnalyzer.Analyze(falseLtObject).Any());
+            var analyzer = GetAnalyzerForRule(containsRule);
+            Assert.IsTrue(analyzer.Analyze(regObj).Any(x => x.Name == RuleName));
         }
 
         [TestMethod]
-        public void VerifyNeqOperator()
+        public void VerifyAnd()
         {
-            var assertTrueObject = new CompareResult()
+            var RuleName = "AndRule";
+            var andRule = new Rule(RuleName)
             {
-                Base = new FileSystemObject("TestPath")
-                {
-                    IsDirectory = true,
-                    Size = 700
-                }
-            };
-
-            var assertFalseObject = new CompareResult()
-            {
-                Base = new FileSystemObject("TestPath2")
-                {
-                    IsDirectory = false,
-                    Size = 701
-                }
-            };
-
-            var stringEquals = new Rule("String Equals Rule")
-            {
+                Expression = "0 AND 1",
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("Path", OPERATION.NEQ)
+                    new Clause("Path", OPERATION.EQ)
                     {
+                        Label = "0",
                         Data = new List<string>()
                         {
-                            "TestPath"
+                            "TestPath1"
                         }
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_TRUE)
+                    {
+                        Label = "1"
                     }
                 }
             };
 
-            var boolEquals = new Rule("Bool Equals Rule")
+            var analyzer = GetAnalyzerForRule(andRule);
+
+            Assert.IsTrue(!analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathOneExecutableObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathTwoExecutableObject).Any(x => x.Name == RuleName));
+        }
+
+        [TestMethod]
+        public void VerifyEmbeddedRulesAreValid()
+        {
+            var analyzer = new Analyzer(AsaHelpers.GetPlatform());
+            Assert.IsTrue(!analyzer.VerifyRules().Any());
+        }
+
+        [TestMethod]
+        public void VerifyInvalidRuleDetection()
+        {
+            var invalidRule = new Rule("Unbalanced Parentheses")
             {
+                Expression = "( 0 AND 1",
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("IsDirectory", OPERATION.NEQ)
+                    new Clause("Path", OPERATION.EQ)
                     {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath2"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.EQ)
+                    {
+                        Label = "1",
                         Data = new List<string>()
                         {
                             "True"
@@ -1049,195 +194,625 @@ namespace AttackSurfaceAnalyzer.Tests
                 }
             };
 
-            var intEquals = new Rule("Int Equals Rule")
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("ClauseInParenthesesLabel")
+            {
+                Expression = "WITH(PARENTHESIS)",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "WITH(PARENTHESIS)"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("CharactersBetweenParentheses")
+            {
+                Expression = "(W(I",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "W(I"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("CharactersBeforeOpenParentheses")
+            {
+                Expression = "W(I",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "W(I"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("CharactersBetweenClosedParentheses")
+            {
+                Expression = "(0 AND W)I)",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "W)I"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("CharactersAfterClosedParentheses")
+            {
+                Expression = "0 AND W)I",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "W)I"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("MultipleConsecutiveNots")
+            {
+                Expression = "0 AND NOT NOT 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "1"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("CloseParenthesesWithNot")
+            {
+                Expression = "(0 AND NOT) 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "1"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("WhiteSpaceLabel")
+            {
+                Expression = "0 AND   ",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("InvalidOperator")
+            {
+                Expression = "0 XAND 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "1"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("InvalidNotOperator")
+            {
+                Expression = "0 NOT AND 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "1"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("EndsWithOperator")
+            {
+                Expression = "0 AND",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("UnusedLabel")
+            {
+                Expression = "0",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "1"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                    {
+                        Label = "0"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("MissingLabel")
             {
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("Size", OPERATION.NEQ)
+                    new Clause("Path", OPERATION.IS_NULL)
                     {
+                        Label = "0"
+                    },
+                    new Clause("Path", OPERATION.IS_NULL)
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("ExpressionRequiresLabels")
+            {
+                Expression = "0 AND 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.IS_NULL),
+                    new Clause("Path", OPERATION.IS_NULL)
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("OutOfOrder")
+            {
+                Expression = "0 1 AND",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
                         Data = new List<string>()
                         {
-                            "700"
+                            "TestPath2"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_TRUE)
+                    {
+                        Label = "1"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+
+            invalidRule = new Rule("StartWithOperator")
+            {
+                Expression = "OR 0 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath2"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.EQ)
+                    {
+                        Label = "1",
+                        Data = new List<string>()
+                        {
+                            "True"
                         }
                     }
                 }
             };
 
-            var boolAnalyzer = GetAnalyzerForRule(boolEquals);
-            var intAnalyzer = GetAnalyzerForRule(intEquals);
-            var stringAnalyzer = GetAnalyzerForRule(stringEquals);
+            Assert.IsFalse(VerifyRule(invalidRule));
 
-            Assert.IsFalse(boolAnalyzer.Analyze(assertTrueObject).Any(x => x.Name == "Bool Equals Rule"));
-            Assert.IsFalse(intAnalyzer.Analyze(assertTrueObject).Any(x => x.Name == "Int Equals Rule"));
-            Assert.IsFalse(stringAnalyzer.Analyze(assertTrueObject).Any(x => x.Name == "String Equals Rule"));
-
-            Assert.IsTrue(boolAnalyzer.Analyze(assertFalseObject).Any(x => x.Name == "Bool Equals Rule"));
-            Assert.IsTrue(intAnalyzer.Analyze(assertFalseObject).Any(x => x.Name == "Int Equals Rule"));
-            Assert.IsTrue(stringAnalyzer.Analyze(assertFalseObject).Any(x => x.Name == "String Equals Rule"));
-        }
-
-        [TestMethod]
-        public void VerifyRegexOperator()
-        {
-            var falseRegexObject = new CompareResult()
+            invalidRule = new Rule("Case Sensitivity")
             {
-                Base = new FileSystemObject("TestPathHere")
-            };
-            var trueRegexObject = new CompareResult()
-            {
-                Base = new FileSystemObject("Directory/File")
-            };
-
-            var regexRule = new Rule("Regex Rule")
-            {
+                Expression = "Variable",
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("Path", OPERATION.REGEX)
+                    new Clause("Path", OPERATION.IS_NULL)
                     {
+                        Label = "VARIABLE"
+                    }
+                }
+            };
+
+            Assert.IsFalse(VerifyRule(invalidRule));
+        }
+
+        [TestMethod]
+        public void VerifyNand()
+        {
+            var RuleName = "NandRule";
+            var nandRule = new Rule(RuleName)
+            {
+                Expression = "0 NAND 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
                         Data = new List<string>()
                         {
-                            ".+\\/.+"
+                            "TestPath1"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "1",
+                    }
+                }
+            };
+
+            var analyzer = GetAnalyzerForRule(nandRule);
+
+            Assert.IsTrue(!analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathOneExecutableObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoExecutableObject).Any(x => x.Name == RuleName));
+        }
+
+        [TestMethod]
+        public void VerifyNor()
+        {
+            var RuleName = "NorRule";
+            var norRule = new Rule(RuleName)
+            {
+                Expression = "0 NOR 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath1"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.EQ)
+                    {
+                        Label = "1",
+                        Data = new List<string>()
+                        {
+                            "True"
                         }
                     }
                 }
             };
 
-            var regexAnalyzer = GetAnalyzerForRule(regexRule);
+            var analyzer = GetAnalyzerForRule(norRule);
 
-            Assert.IsTrue(regexAnalyzer.Analyze(trueRegexObject).Any());
-            Assert.IsFalse(regexAnalyzer.Analyze(falseRegexObject).Any());
+            Assert.IsTrue(!analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathOneExecutableObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathTwoExecutableObject).Any(x => x.Name == RuleName));
         }
 
         [TestMethod]
-        public void VerifyStartsWithOperator()
+        public void VerifyNot()
         {
-            var trueEndsWithObject = new CompareResult()
+            var RuleName = "NotRule";
+            var notRule = new Rule(RuleName)
             {
-                Base = new FileSystemObject("App.exe")
-            };
-            var falseEndsWithObject = new CompareResult()
-            {
-                Base = new FileSystemObject("NotAnApp.pdf")
-            };
-
-            var endsWithRule = new Rule("Ends With Rule")
-            {
+                Expression = "NOT 0",
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("Path", OPERATION.STARTS_WITH)
+                    new Clause("Path", OPERATION.EQ)
                     {
+                        Label = "0",
                         Data = new List<string>()
                         {
-                            "App"
+                            TestPathOne
                         }
                     }
                 }
             };
 
-            var endsWithAnalyzer = GetAnalyzerForRule(endsWithRule);
+            var analyzer = GetAnalyzerForRule(notRule);
 
-            Assert.IsTrue(endsWithAnalyzer.Analyze(trueEndsWithObject).Any());
-            Assert.IsFalse(endsWithAnalyzer.Analyze(falseEndsWithObject).Any());
+            Assert.IsTrue(!analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
         }
 
         [TestMethod]
-        public void VerifyWasModifiedOperator()
+        public void VerifyOr()
         {
-            var falseModifiedObject = new CompareResult()
+            var RuleName = "OrRule";
+            var orRule = new Rule(RuleName)
             {
-                Base = new FileSystemObject("TestPathHere")
-            };
-
-            var alsoFalseModifiedObject = new CompareResult()
-            {
-                Base = new FileSystemObject("TestPathHere"),
-                Compare = new FileSystemObject("TestPathHere")
-                {
-                    IsDirectory = true
-                }
-            };
-
-            var trueModifiedObject = new CompareResult()
-            {
-                Base = new FileSystemObject("Directory/File")
-                {
-                    IsExecutable = true
-                },
-                Compare = new FileSystemObject("Directory/File")
-            };
-
-            var wasModifiedRule = new Rule("Was Modified Rule")
-            {
+                Expression = "0 OR 1",
                 ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("IsExecutable", OPERATION.WAS_MODIFIED)
-                }
-            };
-
-            var regexAnalyzer = GetAnalyzerForRule(wasModifiedRule);
-
-            Assert.IsTrue(regexAnalyzer.Analyze(trueModifiedObject).Any());
-            Assert.IsFalse(regexAnalyzer.Analyze(falseModifiedObject).Any());
-            Assert.IsFalse(regexAnalyzer.Analyze(alsoFalseModifiedObject).Any());
-
-            var trueAlgDict = new CompareResult()
-            {
-                Base = new TpmObject("TestLocal")
-                {
-                    PCRs = new Dictionary<(Tpm2Lib.TpmAlgId, uint), byte[]>()
+                    new Clause("Path", OPERATION.EQ)
                     {
-                        { (TpmAlgId.Sha,1), new byte[5] { 1,2,3,4,5 } }
-                    }
-                },
-                Compare = new TpmObject("TestLocal")
-                {
-                    PCRs = new Dictionary<(Tpm2Lib.TpmAlgId, uint), byte[]>()
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath1"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_TRUE)
                     {
-                        { (TpmAlgId.Sha,1), new byte[5] { 0,0,0,0,0 } }
+                        Label = "1"
                     }
                 }
             };
 
-            var falseAlgDict = new CompareResult()
-            {
-                Base = new TpmObject("TestLocal")
-                {
-                    PCRs = new Dictionary<(Tpm2Lib.TpmAlgId, uint), byte[]>()
-                    {
-                        { (TpmAlgId.Sha,1), new byte[5] { 1,2,3,4,5 } }
-                    }
-                },
-                Compare = new TpmObject("TestLocal")
-                {
-                    PCRs = new Dictionary<(Tpm2Lib.TpmAlgId, uint), byte[]>()
-                    {
-                        { (TpmAlgId.Sha,1), new byte[5] { 1, 2, 3, 4, 5 } }
-                    }
-                }
-            };
+            var analyzer = GetAnalyzerForRule(orRule);
 
-            var pcrsModified = new Rule("Alg Dict Changed PCR 1")
+            Assert.IsTrue(analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathOneExecutableObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoExecutableObject).Any(x => x.Name == RuleName));
+        }
+
+        [TestMethod]
+        public void VerifyValidRuleDetection()
+        {
+            var validRule = new Rule("Regular Rule")
             {
-                ResultType = RESULT_TYPE.TPM,
+                Expression = "0 AND 1",
+                ResultType = RESULT_TYPE.FILE,
                 Flag = ANALYSIS_RESULT_TYPE.FATAL,
                 Clauses = new List<Clause>()
                 {
-                    new Clause("PCRs.(Sha, 1)", OPERATION.WAS_MODIFIED)
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath2"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.EQ)
+                    {
+                        Label = "1",
+                        Data = new List<string>()
+                        {
+                            "True"
+                        }
+                    }
                 }
             };
 
-            var pcrAnalyzer = GetAnalyzerForRule(pcrsModified);
+            Assert.IsTrue(VerifyRule(validRule));
 
-            Assert.IsTrue(pcrAnalyzer.Analyze(trueAlgDict).Any());
-            Assert.IsFalse(pcrAnalyzer.Analyze(falseAlgDict).Any());
+            validRule = new Rule("Extraneous Parenthesis")
+            {
+                Expression = "(0 AND 1)",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath2"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.EQ)
+                    {
+                        Label = "1",
+                        Data = new List<string>()
+                        {
+                            "True"
+                        }
+                    }
+                }
+            };
+
+            Assert.IsTrue(VerifyRule(validRule));
+
+            validRule = new Rule("Deeply Nested Expression")
+            {
+                Expression = "(0 AND 1) OR (2 XOR (3 AND (4 NAND 5)) OR 6)",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            "TestPath2"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.EQ)
+                    {
+                        Label = "1",
+                        Data = new List<string>()
+                        {
+                            "True"
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "2"
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "3"
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "4"
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "5"
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "6"
+                    }
+                }
+            };
+
+            Assert.IsTrue(VerifyRule(validRule));
+
+            validRule = new Rule("StringsForClauseLabels")
+            {
+                Expression = "FOO AND BAR OR BA$_*",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "FOO"
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "BAR"
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_NULL)
+                    {
+                        Label = "BA$_*"
+                    }
+                }
+            };
+
+            Assert.IsTrue(VerifyRule(validRule));
+        }
+
+        [TestMethod]
+        public void VerifyXor()
+        {
+            var RuleName = "XorRule";
+            var xorRule = new Rule(RuleName)
+            {
+                Expression = "0 XOR 1",
+                ResultType = RESULT_TYPE.FILE,
+                Flag = ANALYSIS_RESULT_TYPE.FATAL,
+                Clauses = new List<Clause>()
+                {
+                    new Clause("Path", OPERATION.EQ)
+                    {
+                        Label = "0",
+                        Data = new List<string>()
+                        {
+                            TestPathOne
+                        }
+                    },
+                    new Clause("IsExecutable", OPERATION.IS_TRUE)
+                    {
+                        Label = "1"
+                    }
+                }
+            };
+
+            var analyzer = GetAnalyzerForRule(xorRule);
+
+            Assert.IsTrue(analyzer.Analyze(testPathOneObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathTwoObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(!analyzer.Analyze(testPathOneExecutableObject).Any(x => x.Name == RuleName));
+            Assert.IsTrue(analyzer.Analyze(testPathTwoExecutableObject).Any(x => x.Name == RuleName));
         }
 
         #endregion Public Methods
@@ -1255,6 +830,12 @@ namespace AttackSurfaceAnalyzer.Tests
             };
 
             return new Analyzer(AsaHelpers.GetPlatform(), file);
+        }
+
+        private bool VerifyRule(Rule rule)
+        {
+            var analyzer = GetAnalyzerForRule(rule);
+            return !analyzer.VerifyRules().Any();
         }
 
         #endregion Private Methods
