+++ Tests/Tests.csproj
@@ -57,7 +57,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Lib\Lib.csproj" />
-    <ProjectReference Include="..\Cli\Cli.csproj" />
   </ItemGroup>
   <ItemGroup>
     <Compile Remove="InDatabaseComparatorTests.cs" />
